JetPack 6

+++ pkgs/flash-tools/default.nix
@@ -46,7 +46,7 @@ let
       perl
     ];
 
-    patches = [ ./flash-tools.patch ];
+    patches = [ ./flash-tools-r${lib.versions.major l4tMajorMinorPatchVersion}.patch ];
 
     postPatch = ''
       # Needed in Jetpack 5
@@ -77,12 +77,15 @@ let
       # Wrap x86 binaries in qemu
       pushd bootloader/ >/dev/null
       for filename in chkbdinfo mkbctpart mkbootimg mksparse tegrabct_v2 tegradevflash_v2 tegrahost_v2 tegrakeyhash tegraopenssl tegraparser_v2 tegrarcm_v2 tegrasign_v2; do
-        mv "$filename" ."$filename"-wrapped
-        cat >"$filename" <<EOF
+        if [[ -e $filename ]]; then
+          mv "$filename" ."$filename"-wrapped
+          # DO NOT CHANGE THE WHITESPACE BELOW!
+          cat >"$filename" <<EOF
       #!${runtimeShell}
       exec -a "\$0" ${qemu-user}/bin/qemu-i386 "$out/bootloader/.$filename-wrapped" "\$@"
       EOF
-        chmod +x "$filename"
+          chmod +x "$filename"
+        fi
       done
       popd >/dev/null
     '');
